Move pemToBytes test to its new home.
