+++ test/functional/test-base64.js
@@ -14,7 +14,10 @@
  * limitations under the License.
  */
 
-import {base64UrlDecodeToBytes} from '../../src/base64';
+import {
+  base64UrlDecodeToBytes,
+  pemToBytes,
+} from '../../src/base64';
 
 describe('base64UrlDecodeToBytes', function() {
   it('should map a sample string appropriately', () => {
@@ -23,5 +26,75 @@ describe('base64UrlDecodeToBytes', function() {
     expect(ab[0]).to.equal(1);
     expect(ab[1]).to.equal(0);
     expect(ab[2]).to.equal(1);
+  });
+});
+
+describe('pemToBytes', () => {
+  const PLAIN_TEXT =
+        'MIGfMA0GCSqGSIb3DQEBAQUAA4GNADCBiQKBgQDdlatRjRjogo3WojgGHFHYLugd'
+        + 'UWAY9iR3fy4arWNA1KoS8kVw33cJibXr8bvwUAUparCwlvdbH6dvEOfou0/gCFQs'
+        + 'HUfQrSDv+MuSUMAe8jzKE4qW+jK+xQU9a03GUnKHkkle+Q0pX/g6jXZ7r1/xAK5D'
+        + 'o2kQ+X5xK9cipRgEKwIDAQAB';
+  const PEM = '-----BEGIN PUBLIC KEY-----\n'
+        + 'MIGfMA0GCSqGSIb3DQEBAQUAA4GNADCBiQKBgQDdlatRjRjogo3WojgGHFHYLugd\n'
+        + 'UWAY9iR3fy4arWNA1KoS8kVw33cJibXr8bvwUAUparCwlvdbH6dvEOfou0/gCFQs\n'
+        + 'HUfQrSDv+MuSUMAe8jzKE4qW+jK+xQU9a03GUnKHkkle+Q0pX/g6jXZ7r1/xAK5D\n'
+        + 'o2kQ+X5xK9cipRgEKwIDAQAB\n'
+        + '-----END PUBLIC KEY-----';
+
+  it('should convert a valid key', () => {
+    const binary = pemToBytes(PEM);
+    const plain = atob(PLAIN_TEXT);
+    const len = plain.length;
+    expect(binary.byteLength).to.equal(len);
+    expect(binary[0]).to.equal(plain.charCodeAt(0));
+    expect(binary[1]).to.equal(plain.charCodeAt(1));
+    expect(binary[len - 1]).to.equal(plain.charCodeAt(len - 1));
+    expect(binary[len - 2]).to.equal(plain.charCodeAt(len - 2));
+  });
+
+  it('should convert without headers, footers, line breaks', () => {
+    const binary = pemToBytes(PLAIN_TEXT);
+    const plain = atob(PLAIN_TEXT);
+    const len = plain.length;
+    expect(binary.byteLength).to.equal(len);
+    expect(binary[0]).to.equal(plain.charCodeAt(0));
+    expect(binary[1]).to.equal(plain.charCodeAt(1));
+    expect(binary[len - 1]).to.equal(plain.charCodeAt(len - 1));
+    expect(binary[len - 2]).to.equal(plain.charCodeAt(len - 2));
+  });
+
+  it('should convert without line breaks', () => {
+    const binary = pemToBytes('-----BEGIN PUBLIC KEY-----' + PLAIN_TEXT
+                              + '-----END PUBLIC KEY-----');
+    const plain = atob(PLAIN_TEXT);
+    const len = plain.length;
+    expect(binary.byteLength).to.equal(len);
+    expect(binary[0]).to.equal(plain.charCodeAt(0));
+    expect(binary[1]).to.equal(plain.charCodeAt(1));
+    expect(binary[len - 1]).to.equal(plain.charCodeAt(len - 1));
+    expect(binary[len - 2]).to.equal(plain.charCodeAt(len - 2));
+  });
+
+  it('should convert without header', () => {
+    const binary = pemToBytes(PLAIN_TEXT + '-----END PUBLIC KEY-----');
+    const plain = atob(PLAIN_TEXT);
+    const len = plain.length;
+    expect(binary.byteLength).to.equal(len);
+    expect(binary[0]).to.equal(plain.charCodeAt(0));
+    expect(binary[1]).to.equal(plain.charCodeAt(1));
+    expect(binary[len - 1]).to.equal(plain.charCodeAt(len - 1));
+    expect(binary[len - 2]).to.equal(plain.charCodeAt(len - 2));
+  });
+
+  it('should convert without footer', () => {
+    const binary = pemToBytes('-----BEGIN PUBLIC KEY-----' + PLAIN_TEXT);
+    const plain = atob(PLAIN_TEXT);
+    const len = plain.length;
+    expect(binary.byteLength).to.equal(len);
+    expect(binary[0]).to.equal(plain.charCodeAt(0));
+    expect(binary[1]).to.equal(plain.charCodeAt(1));
+    expect(binary[len - 1]).to.equal(plain.charCodeAt(len - 1));
+    expect(binary[len - 2]).to.equal(plain.charCodeAt(len - 2));
   });
 });
